# Conflicts:
#	src/main/java/com/tang/common/test/BBB.java

+++ src/main/java/com/tang/common/test/BBB.java
@@ -3,7 +3,7 @@ package com.tang.common.test;
 public class BBB {
     public static void main(String[] args) {
 
-        System.out.println(new Integer(300));
+        System.out.println(new Integer(350));
 
         System.out.println(new Integer(200));
     }
